refactor(router): rename routePath to toRoute
